[refactor] : Layout에서 사용되는 SiteFooter 분리
